feat: support init with template
