+++ packages/playground/src/main.ts
@@ -6,29 +6,11 @@ import { getOptions } from './utils';
 import './style.css';
 
 const params = new URLSearchParams(location.search);
-const isAutoInit = params.get('init') !== null;
+const init = params.get('init');
 const isE2E = params.get('room')?.includes('playwright');
 
 const initButton = <HTMLButtonElement>document.getElementById('init-btn');
 const options = getOptions();
-
-/**
- * Manually create initial page structure.
- * In collaboration mode or on page refresh with local persistence,
- * the page structure will be automatically loaded from provider.
- * In these cases, this function should not be called.
- */
-function createInitialPage(workspace: Workspace) {
-  workspace.signals.pageAdded.once(id => {
-    const page = workspace.getPage(id) as Page;
-    const pageBlockId = page.addBlock({ flavour: 'affine:page' });
-    const groupId = page.addBlock({ flavour: 'affine:group' }, pageBlockId);
-    page.addBlock({ flavour: 'affine:paragraph' }, groupId);
-    page.resetHistory();
-  });
-
-  workspace.createPage('page0');
-}
 
 // Subscribe for page update and create editor after page loaded.
 function subscribePage(workspace: Workspace) {
@@ -45,12 +27,10 @@ function subscribePage(workspace: Workspace) {
   });
 }
 
-function main() {
+async function main() {
   const workspace = new Workspace(options).register(BlockSchema);
   // @ts-ignore
   [window.workspace, window.blockSchema] = [workspace, BlockSchema];
-
-  initButton.addEventListener('click', () => createInitialPage(workspace));
 
   // In E2E environment, initial state should be generated by test case,
   // instead of using this default setup.
@@ -58,9 +38,13 @@ function main() {
 
   subscribePage(workspace);
 
-  if (isAutoInit) {
-    createInitialPage(workspace);
-  }
+  const initFunctions = (await import('./data')) as Record<
+    string,
+    (workspace: Workspace) => void
+  >;
+  initButton.addEventListener('click', () => initFunctions.basic(workspace));
+
+  initFunctions[init ?? 'basic']?.(workspace);
 }
 
 main();
